+++ libraries/waspy/wasdi/__init__.py
@@ -42,7 +42,13 @@ Created on 11 Jun 2018
 @author: p.campanella
 """
 from time import sleep
-from __builtin__ import str
+try:
+    from __builtin__ import str
+except Exception as oE0:
+    try:
+        from builtins import str
+    except Exception as oE1:
+        print('Cannot import str. This is bad')
 
 name = "wasdi"
 
@@ -1016,7 +1022,8 @@ def updateProcessStatus(sProcessId, sStatus, iPerc=-1):
                   '  ******************************************************************************')
             return ''
         else:
-            _log('[INFO] waspy.updateProcessStatus: iPerc = -1 - Not considered')
+            # _log('[INFO] waspy.updateProcessStatus: iPerc = -1 - Not considered')
+            pass
     elif iPerc > 100:
         wasdiLog('[ERROR] waspy.updateProcessStatus: iPerc > 100 not valid' +
               '  ******************************************************************************')