+++ wasdiwebserver/src/main/java/it/fadeout/rest/resources/ProductResource.java
@@ -562,7 +562,8 @@ public class ProductResource {
 
 		// Check the file name
 		if (Utils.isNullOrEmpty(sName) || sName.isEmpty()) {
-			sName = "defaultName";
+			//get a random name
+			sName = "defaultName-" + Utils.GetRandomName();
 		}
 
 		// If workspace is not found in DB returns bad request